1/scale is the right thing to accumulate, because
what we are ultimately trying to do is to scale the known count by
the fraction of the daily volume that that is expected to represent.

So we want it to sum to 1 if we are looking at all day, and some
number less than that if we only have part of the day.